Move & Rename DatabaseResolver & it's contract
