+++ src/Affects/Connection/Contracts/ResolvesConnections.php
@@ -1,0 +1,25 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the tenancy/tenancy package.
+ *
+ * Copyright Laravel Tenancy & Daniël Klabbers <daniel@klabbers.email>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ *
+ * @see https://tenancy.dev
+ * @see https://github.com/tenancy
+ */
+
+namespace Tenancy\Affects\Connection\Contracts;
+
+use Tenancy\Identification\Contracts\Tenant;
+use Tenancy\Hooks\Database\Contracts\ProvidesDatabase;
+
+interface ResolvesConnections
+{
+    public function __invoke(Tenant $tenant): ?ProvidesDatabase;
+}
