+++ src/Affects/Connection/composer.json
@@ -1,0 +1,33 @@
+{
+    "name": "tenancy/affects-connection",
+    "description": "The tenancy/tenancy connection mutations",
+    "keywords": ["affects", "tenancy", "connection"],
+    "license": "MIT",
+    "require": {
+        "tenancy/framework": "*"
+    },
+    "autoload": {
+        "psr-4": {
+            "Tenancy\\Affects\\Connection\\": ""
+        }
+    },
+    "authors": [
+        {
+            "name": "Daniël Klabbers",
+            "email": "daniel@klabbers.email",
+            "homepage": "http://luceos.com"
+        },
+        {
+            "name": "Arlon Antonius",
+            "email": "opensource@arlonantonius.com",
+            "homepage": "https://arlonantonius.com"
+        }
+    ],
+    "extra": {
+        "laravel": {
+            "providers": [
+                "Tenancy\\Affects\\Connection\\Provider"
+            ]
+        }
+    }
+}
