+++ src/Tenancy/Providers/Provides/ProvidesListeners.php
@@ -16,14 +16,14 @@ declare(strict_types=1);
 
 namespace Tenancy\Providers\Provides;
 
+use Tenancy\Tenant\Events as Tenant;
 use Illuminate\Support\Facades\Event;
-use Tenancy\Affects\Contracts\ResolvesAffects;
-use Tenancy\Database\Contracts\ResolvesConnections;
 use Tenancy\Database\Events as Database;
 use Tenancy\Database\Listeners as Listen;
 use Tenancy\Identification\Events\Switched;
+use Tenancy\Affects\Contracts\ResolvesAffects;
 use Tenancy\Lifecycle\Contracts\ResolvesHooks;
-use Tenancy\Tenant\Events as Tenant;
+use Tenancy\Affects\Connection\Contracts\ResolvesConnections;
 
 trait ProvidesListeners
 {
