+++ src/Tenancy/Providers/TenancyProvider.php
@@ -20,14 +20,13 @@ use Illuminate\Support\ServiceProvider;
 use Tenancy\Affects\AffectResolver;
 use Tenancy\Affects\Contracts\ResolvesAffects;
 use Tenancy\Database\Contracts\ProvidesPassword;
-use Tenancy\Database\Contracts\ResolvesConnections;
-use Tenancy\Database\DatabaseResolver;
 use Tenancy\Database\PasswordGenerator;
 use Tenancy\Environment;
 use Tenancy\Identification\Contracts\ResolvesTenants;
 use Tenancy\Identification\TenantResolver;
 use Tenancy\Lifecycle\Contracts\ResolvesHooks;
 use Tenancy\Lifecycle\HookResolver;
+use Tenancy\Affects\Connection\Provider;
 
 class TenancyProvider extends ServiceProvider
 {
@@ -40,14 +39,14 @@ class TenancyProvider extends ServiceProvider
         ResolvesHooks::class       => HookResolver::class,
         ResolvesAffects::class     => AffectResolver::class,
         ResolvesTenants::class     => TenantResolver::class,
-        ProvidesPassword::class    => PasswordGenerator::class,
-        ResolvesConnections::class => DatabaseResolver::class,
+        ProvidesPassword::class    => PasswordGenerator::class
     ];
 
     public function register()
     {
         $this->runTrait('register');
 
+        $this->app->register(Provider::class);
         $this->app->register(TenantProvider::class);
     }
